requirements.txt
django==5.2.4
django-environ==0.9.0
django-cors-headers==3.13.0
djangorestframework==3.14.0

Pillow==10.4.0
django-ckeditor==6.7.3
psycopg2==2.9.10


djoser==2.3.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
social-auth-app-django==5.4.2 # <------
social-auth-core==4.5.4 # <------
argon2-cffi==23.1.0

# aws
boto3==1.39.15
django-storages==1.14.6

